Move FundingUtxoError to crate root error module
